upgrade: Renamed mocked routes to match real backend
